+++ Src/FluentAssertions/Equivalency/Steps/EnumerableEquivalencyValidatorExtensions.cs
@@ -14,10 +14,9 @@ internal static class EnumerableEquivalencyValidatorExtensions
             .FailWith(", but found an empty collection.")
             .Then
             .ForCondition(subject.Count == 0 || expectation.Count > 0)
-            .FailWith(", but {0}{2}contains {1} item(s).",
+            .FailWith($", but {{0}}{Environment.NewLine}contains {{1}} item(s).",
                 subject,
-                subject.Count,
-                Environment.NewLine);
+                subject.Count);
     }
 
     public static Continuation AssertCollectionHasEnoughItems<T>(this IAssertionScope scope, ICollection<object> subject,
@@ -25,11 +24,10 @@ internal static class EnumerableEquivalencyValidatorExtensions
     {
         return scope
             .ForCondition(subject.Count >= expectation.Count)
-            .FailWith(", but {0}{3}contains {1} item(s) less than{3}{2}.",
+            .FailWith($", but {{0}}{Environment.NewLine}contains {{1}} item(s) less than{Environment.NewLine}{{2}}.",
                 subject,
                 expectation.Count - subject.Count,
-                expectation,
-                Environment.NewLine);
+                expectation);
     }
 
     public static Continuation AssertCollectionHasNotTooManyItems<T>(this IAssertionScope scope, ICollection<object> subject,
@@ -37,10 +35,9 @@ internal static class EnumerableEquivalencyValidatorExtensions
     {
         return scope
             .ForCondition(subject.Count <= expectation.Count)
-            .FailWith(", but {0}{3}contains {1} item(s) more than{3}{2}.",
+            .FailWith($", but {{0}}{Environment.NewLine}contains {{1}} item(s) more than{Environment.NewLine}{{2}}.",
                 subject,
                 subject.Count - expectation.Count,
-                expectation,
-                Environment.NewLine);
+                expectation);
     }
 }